pyproject.toml
[project]
name = "factaday"
version = "0.1.0"
description = ""
authors = [
    {name = "JeppesS1423",email = "projects@jesperh.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "anthropic (>=0.75.0,<0.76.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
